krew install ctx + ns

+++ fish-install.fish
@@ -18,4 +18,9 @@ end
 # see https://github.com/kubernetes-sigs/krew/pull/811
 # and https://github.com/kubernetes-sigs/krew/issues/810
 $HOME/.krew/bin/kubectl-krew completion fish > $HOME/.config/fish/completions/kubectl-krew.fish
-ln -s $HOME/.krew/bin/kubectl-krew $HOME/.krew/bin/krew
+if test ! -e $HOME/.krew/bin/krew
+  ln -s $HOME/.krew/bin/kubectl-krew $HOME/.krew/bin/krew
+end
+
+grep -v "#" packages/krew.txt > /tmp/krew-packages.txt
+$HOME/.krew/bin/kubectl-krew install < /tmp/krew-packages.txt
